+++ web/src/PersonalContextPanel.test.tsx
@@ -97,6 +97,8 @@ function fakeApi(): Api {
     viewUserConversation: notImplemented,
     getDashboard: notImplemented,
     listAuditLogs: notImplemented,
+    listAiConversations: notImplemented,
+    getAiConversation: notImplemented,
     listConversations: notImplemented,
     getConversation: notImplemented,
     listOperatorOrganizations: notImplemented,